% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRaster.R
\name{testRaster}
\alias{testRaster}
\title{Produces a rasterplot to test event detection tools.}
\usage{
testRaster(length = 60000, size = 1000, activity = 0.5, RNGdata = NULL)
}
\arguments{
\item{length}{Rasterplot length. Defaults to 60,000.}

\item{size}{Number of neurons or input signals. Defaults to 1,000.}

\item{activity}{Activity level from 0 to 1. Defaults to 0.5.}

\item{RNGdata}{A list of seed and RNG kind for random number generation. Defaults to the current workspace configuration.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{neurons}
\keyword{rasterplot}

